refactor: making edits
