+++ src/app/store.ts
@@ -1,18 +1,15 @@
 import { configureStore } from "@reduxjs/toolkit";
 import { setupListeners } from "@reduxjs/toolkit/query";
-import petsReducer from "@/entities/pet/model/slice";
-import shelterReducer from "@/entities/shelter/model/slice";
-import { PetState } from "@entities/pet/index";
-import userReducer from "@/entities/user/model/slice";
-import { UserState } from "@entities/user/index";
-import { ShelterState } from "@entities/shelter/index";
+import * as petModel from "@entities/pet";
+import * as shelterModel from "@entities/shelter";
+import * as userModel from "@entities/user";
 import { baseApi } from "@shared/api";
 export const store = configureStore({
   reducer: {
     [baseApi.reducerPath]: baseApi.reducer,
-    pets: petsReducer,
-    user: userReducer,
-    shelter: shelterReducer,
+    pets: petModel.slice.petsReducer,
+    user: userModel.slice.userReducer,
+    shelter: shelterModel.slice.shelterReducer,
   },
   middleware: (getDefaultMiddleware) =>
     getDefaultMiddleware().concat(baseApi.middleware),
@@ -20,9 +17,9 @@ export const store = configureStore({
 
 setupListeners(store.dispatch);
 export type RootState = {
-  pets: PetState;
-  user: UserState;
-  shelter: ShelterState;
+  pets: petModel.type.PetState;
+  user: userModel.type.UserState;
+  shelter: shelterModel.type.ShelterState;
 };
 
 export type AppDispatch = typeof store.dispatch;
