+++ src/entities/pet/index.ts
@@ -1,5 +1,4 @@
 export { PetCard } from "./ui/PetCard";
-export { default as petsReducer } from "./model/slice";
 export * as slice from "./model/slice";
 export * as api from "./api/api";
 export * as type from "./model/type";
