+++ src/features/pet/searchPet/ui/SearchPet.tsx
@@ -1,5 +1,5 @@
 import React, { ChangeEvent, useState } from "react";
-import * as petModel from "@entities/pet/index";
+import * as petModel from "@entities/pet/";
 import { Input } from "@/shared/ui/input";
 import { Button } from "@shared/ui/button";
 import searchIcon from "@shared/assets/search_icon.svg";
