+++ src/pages/profile/Profile.tsx
@@ -1,8 +1,8 @@
 import React, { useEffect } from "react";
 import { MainContainer } from "@shared/ui/mainContainer";
 import { Navbar } from "@/widgets/navbar";
-import { AddShelter } from "@widgets/addShelter/index.ts";
-import * as petModel from "@entities/pet/index.ts";
+import { AddShelter } from "@widgets/addShelter/";
+import * as petModel from "@entities/pet/";
 export const Profile: React.FC = () => {
   const { data: favorites } = petModel.api.useGetFavoritesQuery();
   useEffect(() => {}, [favorites]);
